single fucntion for generateShortPDF
